Revert "routing rules: mark support"
This reverts commit ee88c71d16a6b777c916275401a971eea78871d6.

+++ include/netlink-types.h
@@ -282,7 +282,7 @@ struct rtnl_rule
 {
 	NLHDR_COMMON
 
-	uint32_t	r_mark;
+	uint64_t	r_mark;
 	uint32_t	r_prio;
 	uint32_t	r_realms;
 	uint32_t	r_table;
